Project 2 - Breakout strategy
"In this project, you will code and evaluate a breakout signal. You will
run statistical tests to test for normality and to find alpha. You will
also learn to find outliers and evaluate the effect that filtered outliers
could have on your trading signal. You will run various scenarios of
your model with or without the outliers and decide if the outliers
should be kept or not."