Sentiment Analysis using NLP
"In this project, you will apply Natural Language Processing on
corporate filings, such as 10Q and 10K statements, from cleaning
data and text processing, to feature extraction and modeling. You
will utilize bag-of-words and TF-IDF to generate company-specific
sentiments. Based on the sentiments, you will decide which company
to invest in, and the optimal time to buy or sell."